Fix getDefaultConfig usage
